AFVU
Remote bitcoin mining management software
SmartCoin - Simple Miner Administration for Remote Terminals As the name implies, this is a full-blown administration system for miners that is specifically designed with terminals in mind (remote ssh shell access, etc.). It is also so much more - its a full load-balanced system that tracks all of the miner instances on a machine, and reports everything on one easy to view screen. The idea is to spread many miner instances across several GPUs for multiple pools (and even solo!) on one machine (i'm mining to 4 concurrent pools as we speak. If any one of them goes down, the system re-balances its self with any remaining good pools, and my hashing power remains the same). SmartCoin also makes automated deployments a breeze - it can start everything up (according to "mapping rules") at boot up, and makes it simple to restart an entire array of miners. You can even change settings on the fly and watch as the system magically adjusts itsself. SmartCoin is fully backed by a MySql database, and written in 100% linux CLI tools (bash, sed, awk). Another neat perk to this, is that you can put a small percentage of your hashing power into solo mining if you like, while you're mining to several pools at once! This is NOT a proxy - its a load balancing system that does not interfere with Long Polling or suffer from any other problems experienced with proxies.